Add files via upload
CreepyPasta Library!

Goals:

Create the User and Pasta models with all appropriate relationships
Incorporate a validated login/registration page
On the main page, allow the user to add a new Pasta, with validations. Added Pastas should automatically be favorited by the logged-in user.
Have a list of all the CreepyPastas on the main page displaying the title and the user who uploaded the book
When the title is clicked, display a page with the Pasta's information, including a list of all users who have favorited that Pasta.
If the logged in user has favorited the Pasta, they should be able to "un-favorite"it as well.
If the logged in user has not yet favorited the Pasta, there should be a link to add this Pasta to their favorites
If the logged in user is the uploader of the Pasta, it should allow them to edit (same validations apply) or delete the Pasta, when clicking on viewing it.